fixed TurboSrcButtonOpen refactor
